feat(clustering) factor some protobuf generic functions
