+++ kong/db/declarative/init.lua
@@ -565,6 +565,8 @@ local function remove_nulls(tbl)
   return tbl
 end
 
+declarative.remove_nulls = remove_nulls
+
 
 function declarative.get_current_hash()
   return lmdb.get(DECLARATIVE_HASH_KEY)
