+++ kong/tools/wrpc.lua
@@ -17,100 +17,6 @@ local CHANNEL_CLIENT_PREFIX = "wrpc_client_"
 
 local wrpc = {}
 
-do
-  local structpb_value, structpb_list, structpb_struct
-
-  function structpb_value(v)
-    local t = type(v)
-
-    local bool_v = nil
-    if t == "boolean" then
-      bool_v = v
-    end
-
-    local list_v = nil
-    local struct_v = nil
-
-    if t == "table" then
-      if t[1] ~= nil then
-        list_v = structpb_list(v)
-      else
-        struct_v = structpb_struct(v)
-      end
-    end
-
-    return {
-      null_value = t == "nil" and 1 or nil,
-      bool_value = bool_v,
-      number_value = t == "number" and v or nil,
-      string_value = t == "string" and v or nil,
-      list_value = list_v,
-      struct_value = struct_v,
-    }
-  end
-
-  function structpb_list(l)
-    local out = {}
-    for i, v in ipairs(l) do
-      out[i] = structpb_value(v)
-    end
-    return { values = out }
-  end
-
-  function structpb_struct(d)
-    local out = {}
-    for k, v in pairs(d) do
-      out[k] = structpb_value(v)
-    end
-    return { fields = out }
-  end
-
-  wrpc.pbwrap_struct = structpb_struct
-end
-
-do
-  local structpb_value, structpb_list, structpb_struct
-
-  function structpb_value(v)
-    if type(v) ~= "table" then
-      return v
-    end
-
-    if v.list_value then
-      return structpb_list(v.list_value)
-    end
-
-    if v.struct_value then
-      return structpb_struct(v.struct_value)
-    end
-
-    return v.bool_value or v.string_value or v.number_value or v.null_value
-  end
-
-  function structpb_list(l)
-    local out = {}
-    if type(l) == "table" then
-      for i, v in ipairs(l.values or l) do
-        out[i] = structpb_value(v)
-      end
-    end
-    return out
-  end
-
-  function structpb_struct(struct)
-    if type(struct) ~= "table" then
-      return struct
-    end
-
-    local out = {}
-    for k, v in pairs(struct.fields or struct) do
-      out[k] = structpb_value(v)
-    end
-    return out
-  end
-
-  wrpc.pbunwrap_struct = structpb_struct
-end
 
 local Queue = {}
 Queue.__index = Queue
@@ -339,6 +245,10 @@ local wrpc_peer = {
 }
 wrpc_peer.__index = wrpc_peer
 
+local function is_wsclient(conn)
+  return conn and not conn.close or nil
+end
+
 --- a `peer` object holds a (websocket) connection and a service.
 function wrpc.new_peer(conn, service, opts)
   opts = opts or {}
@@ -346,7 +256,7 @@ function wrpc.new_peer(conn, service, opts)
     conn = conn,
     service = service,
     seq = 1,
-    request_queue = (conn and not conn.close) and Queue.new(),
+    request_queue = is_wsclient(conn) and Queue.new(),
     response_queue = {},
     closing = false,
     channel_dict = opts.channel and ngx.shared[CHANNEL_DICT_NAME],
@@ -523,17 +433,19 @@ function wrpc_peer:handle(payload)
       local input_data = decodearray(self.decode, rpc.input_type, payload.payloads)
       local ok, output_data = ok_wrapper(pcall(rpc.handler, self, table_unpack(input_data, 1, input_data.n)))
       if not ok then
+        local err = tostring(output_data[1])
+        ngx.log(ngx.ERR, ("[wrpc] Error handling %q method: %q"):format(rpc.name, err))
         self:send_payload({
           mtype = "MESSAGE_TYPE_ERROR",
           error = {
             etype = "ERROR_TYPE_UNSPECIFIED",
-            description = tostring(output_data),
+            description = err,
           },
           srvc_id = payload.svc_id,
           rpc_id = payload.rpc_id,
           ack = payload.seq,
         })
-        return nil, output_data
+        return nil, err
       end
 
       self:send_payload({
